Interval estimators

This is the parent class for all confidence intervals implemented in this package. Currently, only confidence intervals for the parameter \(\mu\) of a normal distribution are implemented. Details about the methods for calculating confidence intervals can be found in (our upcoming paper).

Usage

IntervalEstimator(two_sided, l1, u1, l2, u2, label)

RepeatedCI(two_sided = TRUE)

StagewiseCombinationFunctionOrderingCI(two_sided = TRUE)

MLEOrderingCI(two_sided = TRUE)

LikelihoodRatioOrderingCI(two_sided = TRUE)

ScoreTestOrderingCI(two_sided = TRUE)

NeymanPearsonOrderingCI(two_sided = TRUE, mu0 = 0, mu1 = 0.4)

NaiveCI(two_sided = TRUE)

Arguments

two_sided

logical indicating whether the confidence interval is two-sided.

l1

functional representation of the lower boundary of the interval in the early futility and efficacy regions.

u1

functional representation of the upper boundary of the interval in the early futility and efficacy regions.

l2

functional representation of the lower boundary of the interval in the continuation region.

u2

functional representation of the upper boundary of the interval in the continuation region.

label

name of the estimator. Used in printing methods.

mu0

expected value of the normal distribution under the null hypothesis.

mu1

expected value of the normal distribution under the null hypothesis.

Value

an object of class IntervalEstimator.

Details

The implemented confidence intervals are:

MLEOrderingCI()

LikelihoodRatioOrderingCI()

ScoreTestOrderingCI()

StagewiseCombinationFunctionOrderingCI()

These confidence intervals are constructed by specifying an ordering of the sample space and finding the value of \(\mu\), such that the observed sample is the \(\alpha/2\) (or (\(1-\alpha/2\))) quantile of the sample space according to the chosen ordering. Some of the implemented orderings are based on the work presented in (Emerson and Fleming 1990) , (Sections 8.4 in Jennison and Turnbull 1999) , and (Sections 4.1.1 and 8.2.1 in Wassmer and Brannath 2016) .

References

Emerson SS, Fleming TR (1990). “Parameter estimation following group sequential hypothesis testing.” Biometrika, 77(4), 875--892. doi:10.2307/2337110 .

Jennison C, Turnbull BW (1999). Group Sequential Methods with Applications to Clinical Trials, 1 edition. Chapman and Hall/CRC., New York. doi:10.1201/9780367805326 .

Wassmer G, Brannath W (2016). Group Sequential and Confirmatory Adaptive Designs in Clinical Trials, 1 edition. Springer, Cham, Switzerland. doi:10.1007/978-3-319-32562-0 .

See also

Examples

# This is the definition of the 'naive' confidence interval for one-armed trials
IntervalEstimator(
  two_sided = TRUE,
  l1 = \(smean1, n1, sigma, ...) smean1 - qnorm(.95, sd = sigma/sqrt(n1)),
  u1 = \(smean1, n1, sigma, ...) smean1 + qnorm(.95, sd = sigma/sqrt(n1)),
  l2 = \(smean1, smean2, n1, n2, sigma, ...) smean2 - qnorm(.95, sd = sigma/sqrt(n1 + n2)),
  u2 = \(smean1, smean2, n1, n2, sigma, ...) smean2 + qnorm(.95, sd = sigma/sqrt(n1 + n2)),
  label="My custom CI")
#> My custom CI