Point estimators

This is the parent class for all point estimators implemented in this package. Currently, only estimators for the parameter \(\mu\) of a normal distribution are implemented.

Usage

PointEstimator(g1, g2, label)

SampleMean()

FirstStageSampleMean()

WeightedSampleMean(w1 = 0.5)

AdaptivelyWeightedSampleMean(w1 = 1/sqrt(2))

MinimizePeakVariance()

BiasReduced(iterations = 1L)

RaoBlackwell()

PseudoRaoBlackwell()

MidpointStagewiseCombinationFunctionOrderingCI()

MidpointMLEOrderingCI()

MidpointLikelihoodRatioOrderingCI()

MidpointScoreTestOrderingCI()

MidpointNeymanPearsonOrderingCI()

MedianUnbiasedStagewiseCombinationFunctionOrdering()

MedianUnbiasedMLEOrdering()

MedianUnbiasedLikelihoodRatioOrdering()

MedianUnbiasedScoreTestOrdering()

MedianUnbiasedNeymanPearsonOrdering(mu0 = 0, mu1 = 0.4)

Arguments

g1

functional representation of the estimator in the early futility and efficacy regions.

g2

functional representation of the estimator in the continuation region.

label

name of the estimator. Used in printing methods.

w1

weight of the first-stage data.

iterations

number of bias reduction iterations. Defaults to 1.

mu0

expected value of the normal distribution under the null hypothesis.

mu1

expected value of the normal distribution under the null hypothesis.

Value

an object of class PointEstimator. This class signals that an object can be supplied to the evaluate_estimator and the analyze functions.

Details

Details about the point estimators can be found in (our upcoming paper).

Sample Mean (SampleMean())

The sample mean is the maximum likelihood estimator for the mean and probably the 'most straightforward' of the implemented estimators.

Fixed weighted sample means (WeightedSampleMean())

The first- and second-stage (if available) sample means are combined via fixed, predefined weights. See (Brannath et al. 2006) and (Section 8.3.2 in Wassmer and Brannath 2016) .

Adaptively weighted sample means (AdaptivelyWeightedSampleMean())

The first- and second-stage (if available) sample means are combined via a combination of fixed and adaptively modified weights that depend on the standard error. See (Section 8.3.4 in Wassmer and Brannath 2016) .

Minimizing peak variance in adaptively weighted sample means (MinimizePeakVariance())

For this estimator, the weights of the adaptively weighted sample mean are chosen to minimize the variance of the estimator for the value of \(\mu\) which maximizes the expected sample size.

(Pseudo) Rao-Blackwell estimators (RaoBlackwell and PseudoRaoBlackwell)

The conditional expectation of the first-stage sample mean given the overall sample mean and the second-stage sample size. See (Emerson and Kittelson 1997) .

A bias-reduced estimator (BiasReduced())

This estimator is calculated by subtracting an estimate of the bias from the MLE. See (Whitehead 1986) .

Median-unbiased estimators

The implemented median-unbiased estimators are:

MedianUnbiasedMLEOrdering()

MedianUnbiasedLikelihoodRatioOrdering()

MedianUnbiasedScoreTestOrdering()

MedianUnbiasedStagewiseCombinationFunctionOrdering()

These estimators are constructed by specifying an ordering of the sample space and finding the value of \(\mu\), such that the observed sample is the median of the sample space according to the chosen ordering. Some of the implemented orderings are based on the work presented in (Emerson and Fleming 1990) , (Sections 8.4 in Jennison and Turnbull 1999) , and (Sections 4.1.1 and 8.2.1 in Wassmer and Brannath 2016) .

References

Brannath W, König F, Bauer P (2006). “Estimation in flexible two stage designs.” Statistics in Medicine, 25(19), 3366-3381. doi:10.1002/sim.2258 .

Emerson SS, Fleming TR (1990). “Parameter estimation following group sequential hypothesis testing.” Biometrika, 77(4), 875--892. doi:10.2307/2337110 .

Emerson SS, Kittelson JM (1997). “A computationally simpler algorithm for the UMVUE of a normal mean following a group sequential trial.” Biometrics, 53(1), 365--369. doi:10.2307/2533122 .

Jennison C, Turnbull BW (1999). Group Sequential Methods with Applications to Clinical Trials, 1 edition. Chapman and Hall/CRC., New York. doi:10.1201/9780367805326 .

Wassmer G, Brannath W (2016). Group Sequential and Confirmatory Adaptive Designs in Clinical Trials, 1 edition. Springer, Cham, Switzerland. doi:10.1007/978-3-319-32562-0 .

Whitehead J (1986). “On the bias of maximum likelihood estimation following a sequential test.” Biometrika, 73(3), 573--581. doi:10.2307/2336521 .

See also

Examples

PointEstimator(g1 = \(smean1, ...) smean1,g2 = \(smean2, ...) smean2, label="My custom estimator")
#> My custom estimator