Statistics and Estimators of the adestr package

The Statistic class is a parent class for the classes Estimator and PValue. The Estimator class is a parent for the classes PointEstimator and ConfidenceInterval.

Arguments

label

name of the statistic. Used in printing methods.

Value

An object of class Statistic. This class signals that an object can be supplied to the analyze function.

Details

The function analyze can be used to calculate the value of a Statistic for a given dataset.

The function evaluate_estimator can be used to evaluate distributional quantities of an Estimator like the MSE for a PointEstimator or the Coverage for a ConfidenceInterval.

See also