Generate an exemplary adaptive design

The design was optimized to minimize the expected sample size under the alternative hypothesis for a one-armed trial. The boundaries are chosen to control the type I error at 0.025 for a normally distributed test statistic (i.e. known variance). For an alternative hypothesis of mu=0.4, the overall power is 80%.

Usage

get_example_design(two_armed = FALSE, label = NULL)

Arguments

two_armed

(logical) determins whether the design is for one- or two-armed trials.

label

(optional) label to be assigned to the design.

Value

an exemplary design of class TwoStageDesign. This object contains information about the sample size recalculation rule n2, the futility and efficacy boundaries c1f and c1e and the second-stage rejection boundary c2.

Examples

get_example_design()
#> TwoStageDesign<n1=28;0.8<=x1<=2.3:n2=9-40>