adestr

Point estimates, confidence intervals, and p-values for optimal adaptive two-stage designs.

Details

This package implements methods to evaluate the performance characteristics of various point and interval estimators for optimal adaptive two-stage designs. Specifically, this package is written to interface with trial designs created by the adoptr package (Kunzmann et al. 2021; Pilz et al. 2021) . Apart from the a priori evaluation of performance characteristics, this package also allows for the calculation of the values of the estimators given real datasets, and it implements methods to calculate p-values.

References

Kunzmann K, Pilz M, Herrmann C, Rauch G, Kieser M (2021). “The adoptr package: Adaptive Optimal Designs for Clinical Trials in R.” Journal of Statistical Software, 98(9), 1–21. doi:10.18637/jss.v098.i09 .

Pilz M, Kunzmann K, Herrmann C, Rauch G, Kieser M (2021). “Optimal planning of adaptive two-stage designs.” Statistics in Medicine, 40(13), 3196-3213. doi:10.1002/sim.8953 .

See also

Author

Maintainer: Jan Meis meis@imbi.uni-heidelberg.de (ORCID)

Other contributors:

Martin Maechler maechler@stat.math.ethz.ch (ORCID) (Original author of monoSpl.c (from the 'stats' package).) [copyright holder]